Create Variable Fonts
Variable fonts are the built-in latest upfront. It’s a single file that comprises every version of a font that a user will need to view the design. While there are only a few lists of variable fonts to work with, it is growing, and this is where we are going with type on the web.

To implement variable fonts, we must select a font that supports the feature and a browser that has implemented the font-variation-settings property.